refactor(mps-sync-plugin-lib): move change handlers from ModelixTreeChangeListener
to the corresponding Transformer classes
