+++ mps-sync-plugin-lib/src/main/kotlin/org/modelix/mps/sync/transformation/modelixToMps/transformers/ModuleTransformer.kt
@@ -16,21 +16,32 @@
 
 package org.modelix.mps.sync.transformation.modelixToMps.transformers
 
+import com.intellij.openapi.diagnostic.logger
+import jetbrains.mps.model.ModelDeleteHelper
+import jetbrains.mps.module.ModuleDeleteHelper
 import jetbrains.mps.project.AbstractModule
 import jetbrains.mps.project.MPSProject
 import jetbrains.mps.project.ModuleId
+import jetbrains.mps.project.Project
 import jetbrains.mps.project.structure.modules.ModuleReference
+import jetbrains.mps.project.structure.modules.SolutionDescriptor
+import jetbrains.mps.refactoring.Renamer
+import org.jetbrains.mps.openapi.module.SModule
 import org.jetbrains.mps.openapi.module.SModuleId
 import org.jetbrains.mps.openapi.persistence.PersistenceFacade
 import org.modelix.kotlin.utils.UnstableModelixFeature
 import org.modelix.model.api.BuiltinLanguages
 import org.modelix.model.api.INode
+import org.modelix.mps.sync.mps.ActiveMpsProjectInjector
 import org.modelix.mps.sync.mps.factories.SolutionProducer
 import org.modelix.mps.sync.tasks.SyncDirection
 import org.modelix.mps.sync.tasks.SyncLock
 import org.modelix.mps.sync.tasks.SyncQueue
+import org.modelix.mps.sync.transformation.cache.ModuleWithModuleReference
 import org.modelix.mps.sync.transformation.cache.MpsToModelixMap
+import org.modelix.mps.sync.util.BooleanUtil
 import org.modelix.mps.sync.util.nodeIdAsLong
+import java.text.ParseException
 
 @UnstableModelixFeature(reason = "The new modelix MPS plugin is under construction", intendedFinalization = "2024.1")
 class ModuleTransformer(private val nodeMap: MpsToModelixMap, private val syncQueue: SyncQueue, project: MPSProject) {
@@ -41,6 +52,8 @@ class ModuleTransformer(private val nodeMap: MpsToModelixMap, private val syncQu
             return PersistenceFacade.getInstance().createModuleId(uuid)
         }
     }
+
+    private val logger = logger<ModuleTransformer>()
 
     private val solutionProducer = SolutionProducer(project)
 
@@ -78,5 +91,81 @@ class ModuleTransformer(private val nodeMap: MpsToModelixMap, private val syncQu
         }
 
         nodeMap.put(parentModule, moduleReference, iNode.nodeIdAsLong())
+    }
+
+    fun modulePropertyChanged(role: String, nodeId: Long, sModule: SModule, newValue: String?) {
+        val moduleId = sModule.moduleId
+        if (sModule !is AbstractModule) {
+            logger.error("SModule ($moduleId) is not an AbstractModule, therefore its $role property cannot be changed. Corresponding Modelix Node ID is $nodeId.")
+            return
+        }
+
+        if (role == BuiltinLanguages.jetbrains_mps_lang_core.INamedConcept.name.getSimpleName()) {
+            val oldValue = sModule.moduleName
+            if (oldValue != newValue) {
+                if (newValue.isNullOrEmpty()) {
+                    logger.error("Name cannot be null or empty for Module $moduleId. Corresponding Modelix Node ID is $nodeId.")
+                    return
+                }
+
+                val activeProject = ActiveMpsProjectInjector.activeMpsProject as Project
+                Renamer(activeProject).renameModule(sModule, newValue)
+            }
+        } else if (role == BuiltinLanguages.MPSRepositoryConcepts.Module.moduleVersion.getSimpleName()) {
+            try {
+                val newVersion = newValue?.toInt() ?: return
+                val oldVersion = sModule.moduleVersion
+                if (oldVersion != newVersion) {
+                    sModule.moduleVersion = newVersion
+                }
+            } catch (ex: NumberFormatException) {
+                logger.error("New module version ($newValue) of SModule ($moduleId) is not an integer, therefore it cannot be set in MPS. Corresponding Modelix Node ID is $nodeId.")
+            }
+        } else if (role == BuiltinLanguages.MPSRepositoryConcepts.Module.compileInMPS.getSimpleName()) {
+            try {
+                val newCompileInMPS = newValue?.let { BooleanUtil.toBooleanStrict(it) } ?: return
+                val moduleDescriptor = sModule.moduleDescriptor ?: return
+                val oldCompileInMPS = moduleDescriptor.compileInMPS
+                if (oldCompileInMPS != newCompileInMPS) {
+                    if (moduleDescriptor !is SolutionDescriptor) {
+                        logger.error("Module ($moduleId)'s descriptor is not a SolutionDescriptor, therefore compileInMPS will not be (un)set in MPS. Corresponding Modelix Node ID is $nodeId.")
+                        return
+                    }
+                    moduleDescriptor.compileInMPS = newCompileInMPS
+                }
+            } catch (ex: ParseException) {
+                logger.error("New compileInMPS ($newValue) property of SModule ($moduleId) is not a strict boolean, therefore it cannot be set in MPS. Corresponding Modelix Node ID is $nodeId.")
+            }
+        } else {
+            logger.error("Role $role is unknown for concept Module. Therefore the property is not set in MPS from Modelix Node $nodeId")
+        }
+    }
+
+    fun moduleDeleted(sModule: SModule, nodeId: Long) {
+        sModule.models.forEach { model ->
+            val modelNodeId = nodeMap[model]
+            ModelDeleteHelper(model).delete()
+            modelNodeId?.let { nodeMap.remove(it) }
+        }
+        val project = ActiveMpsProjectInjector.activeMpsProject!!
+        ModuleDeleteHelper(project).deleteModules(listOf(sModule), false, true)
+        nodeMap.remove(nodeId)
+    }
+
+    fun outgoingModuleReferenceFromModuleDeleted(moduleWithModuleReference: ModuleWithModuleReference, nodeId: Long) {
+        val sourceModule = moduleWithModuleReference.source
+        if (sourceModule !is AbstractModule) {
+            logger.error("Source module ($sourceModule) is not an AbstractModule, therefore outgoing module dependency reference cannot be removed. Corresponding Modelix Node ID is $nodeId.")
+            return
+        }
+
+        val targetModuleReference = moduleWithModuleReference.moduleReference
+        val dependency =
+            sourceModule.moduleDescriptor?.dependencies?.firstOrNull { it.moduleRef == targetModuleReference }
+        if (dependency != null) {
+            sourceModule.removeDependency(dependency)
+        } else {
+            logger.error("Outgoing dependency $targetModuleReference from Module $sourceModule is not found, therefore it cannot be deleted. Corresponding Modelix Node ID is $nodeId.")
+        }
     }
 }
